Add certificate validator
