+++ src/Shark.Fido2.Metadata.Core/DependencyInjection.cs
@@ -7,6 +7,7 @@ public static class DependencyInjection
 {
     public static void RegisterMetadataService(this IServiceCollection services)
     {
+        services.AddTransient<ICertificateValidator, CertificateValidator>();
         services.AddTransient<IMetadataService, MetadataService>();
     }
 }
